refactor: user item actions
