+++ frontend/src/components/Users/UsersList/UserItem/UserItem.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 
+import Badge from '@/components/Primitives/Badge/Badge';
 import Icon from '@/components/Primitives/Icons/Icon/Icon';
 import Flex from '@/components/Primitives/Layout/Flex/Flex';
 import Text from '@/components/Primitives/Text/Text';
@@ -8,8 +9,7 @@ import useCurrentSession from '@/hooks/useCurrentSession';
 import { InnerContainer } from '@/styles/pages/pages.styles';
 import { UserWithTeams } from '@/types/user/user';
 
-import CardEnd from '../partials/CardUser/CardEnd';
-import SuperAdmin from '../partials/CardUser/SuperAdmin';
+import UserItemActions from './UserItemActions/UserItemActions';
 import UserTitle from './UserTitle/UserTitle';
 
 type UserItemProps = {
@@ -17,17 +17,14 @@ type UserItemProps = {
 };
 
 const UserItem = React.memo<UserItemProps>(({ userWithTeams }) => {
-  const { userId, isSAdmin } = useCurrentSession();
-
-  const { email, isSAdmin: isUserSAdmin, _id } = userWithTeams.user;
-  const { teamsNames } = userWithTeams;
-  const { user } = userWithTeams;
+  const { isSAdmin } = useCurrentSession();
+  const { teamsNames, user } = userWithTeams;
 
   const getTeamsCountText = () => {
     if (teamsNames?.length === 1) {
       return 'in 1 team';
     }
-    if (teamsNames?.length !== 0 && teamsNames?.length !== 1 && teamsNames) {
+    if (teamsNames && teamsNames?.length > 1) {
       return `in ${teamsNames.length} teams`;
     }
     return 'no teams';
@@ -53,45 +50,31 @@ const UserItem = React.memo<UserItemProps>(({ userWithTeams }) => {
 
         <Flex align="center" css={{ flex: '1' }}>
           <Text color="primary300" size="sm">
-            {email}
+            {user.email}
           </Text>
         </Flex>
 
-        {!isSAdmin && (
-          <Flex align="center" css={{ flex: '1' }}>
-            <SuperAdmin
-              userSAdmin={isUserSAdmin}
-              loggedUserSAdmin={isSAdmin}
-              userId={_id}
-              loggedUserId={userId}
-            />
-          </Flex>
-        )}
-
-        <Flex align="center" css={{ justifyContent: 'end', width: '$683' }} gap="8">
-          <Flex align="center" css={{ ml: '$40', alignItems: 'center' }} gap="8">
-            <Flex align="center" css={{ width: '$147' }}>
-              <Tooltip content={teamsSeparatedByComma}>
-                <Text css={{ mr: '$2', fontWeight: '$bold', cursor: 'default' }} size="sm">
-                  {getTeamsCountText()}
-                </Text>
-              </Tooltip>
-            </Flex>
-          </Flex>
-          {isSAdmin && (
-            <Flex css={{ width: '40%' }} justify="end">
-              <Flex align="center" css={{ width: '$237' }} justify="start">
-                <SuperAdmin
-                  userSAdmin={isUserSAdmin}
-                  loggedUserSAdmin={isSAdmin}
-                  userId={_id}
-                  loggedUserId={userId}
-                />
-              </Flex>
-              <CardEnd user={userWithTeams.user} />
-            </Flex>
+        <Flex align="center" css={{ flex: '1' }}>
+          {!isSAdmin && user.isSAdmin && (
+            <Badge pill variant="success" size="sm">
+              SUPER ADMIN
+            </Badge>
           )}
         </Flex>
+
+        <Flex align="center" justify="end" css={{ flex: '1' }}>
+          <Tooltip content={teamsSeparatedByComma}>
+            <Text css={{ cursor: 'default' }} fontWeight="bold" size="sm">
+              {getTeamsCountText()}
+            </Text>
+          </Tooltip>
+        </Flex>
+
+        {isSAdmin && (
+          <Flex align="center" justify="end" css={{ flex: '1' }}>
+            <UserItemActions user={user} />
+          </Flex>
+        )}
       </InnerContainer>
     </Flex>
   );
